+++ src/aks-preview/azext_aks_preview/custom.py
@@ -1999,6 +1999,7 @@ def aks_agentpool_update(
     allowed_host_ports=None,
     asg_ids=None,
     enable_artifact_streaming=False,
+    disable_artifact_streaming=False,
     enable_managed_gpu=False,
     os_sku=None,
     ssh_access=None,